Finished achievements
Hell drop anim finished and polished!
Added weapon drop for all weapons.
Now, the stratagemFMW will check if a point is already targeted for a hell pod and not allow to target it.
Items: it'll display that the item was destroyed when an enemy goes on its tile (and the enemy won't acquire a weapon anymore!)
WIP m43 machin gun logic for the 3rd shot.