Finished deployables
Added a new drop anim (with the guiding laser)
New Delivery icon!
Improved Mg43's icon.
Improved mechs' sprites
Improving stratagemFMW tip image (still has an issue with the end of the second cycle of the tip image)
Attempt to do a tip image for the passive, but I didn't managed to display some (fake) logic.
Increased the core cost of the damage upgrade of Patriot's weapons (from 2 to 3)